fix: breakout map into its own component
